Code refactor and helper functions folder added
