+++ src/components/App.scss
@@ -19,4 +19,10 @@
   flex-wrap: wrap;
   align-content: flex-start;
   }
+
+  .search {
+    display: flex;
+    padding: 10px;
+    margin: 15px 10px 10px 5px;
+  }
   }
